refactor: Divide Chip and ChipGroup samples so it becomes easier to load the samples pages
