+++ Uno.Gallery/Uno.Gallery.UWP/Views/Styles/ScrollViewer.xaml
@@ -1,0 +1,16 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+					xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+	<Style x:Key="HorizontalScrollViewerStyle"
+		   TargetType="ScrollViewer">
+		<Setter Property="HorizontalScrollMode"
+				Value="Auto" />
+		<Setter Property="HorizontalScrollBarVisibility"
+				Value="Auto" />
+		<Setter Property="VerticalScrollMode"
+				Value="Disabled" />
+		<Setter Property="VerticalScrollBarVisibility"
+				Value="Hidden" />
+	</Style>
+
+</ResourceDictionary>
